Conflicts:
	momo_modules/pic12_executive/src/bootloader.h
	momo_modules/pic12_executive/src/i2c_defines.h

+++ tools/automation/provision.sh
@@ -1,28 +1,39 @@
 #!/bin/bash
 
 die() { echo "$@" 1>&2; exit 1; }
+set -e
+
+echo "MOMO_DEV: $MOMO_DEV"
 
 apt-get update
-apt-get install -y python python-setuptools python-dev libc6:i386 lib32stdc++6 gpsim
+apt-get install -y python python-setuptools python-dev python-pip
+if [ -n "$MOMO_DEV" ]; then
+	apt-get install -y libc6:i386 lib32stdc++6
+fi
+apt-get install wget
 
-easy_install pip
-easy_install intelhex # This doesn't work with PIP for some reason
-pip install cmdln ZODB3 colorama pyparsing BeautifulSoup4 Cheetah pyserial pytest decorator pycparser nose
-pip install --allow-external dirspec --allow-unverified dirspec dirspec
+easy_install -U pip
+
+pip install pyserial
+pip install --allow-external intelhex --allow-unverified intelhex pymomo
+
+pip install http://sourceforge.net/projects/scons/files/latest/download --egg | tee -a $HOME/scons-install.log
+PYTHONPATH=`cat $HOME/scons-install.log | grep ' library modules ' | awk '{print $6}'`
+echo "PYTHONPATH=$PYTHONPATH" >> $HOME/.profile
 
 if [ -n "$TRAVIS" ]; then
-	MOMOROOT=`pwd`
-	DOWNLOADCACHE="~/.cached_downloads"
+	export MOMOPATH=`pwd`
+	export DOWNLOADCACHE="~/.cached_downloads"
 else # VAGRANT
-	MOMOROOT="/vagrant"
-	HOME="/home/vagrant"
-	DOWNLOADCACHE="/vagrant/.cached_downloads"
+	export MOMOPATH="/vagrant"
+	export HOME="/home/vagrant"
+	export DOWNLOADCACHE="/vagrant/.cached_downloads"
 
 	echo "Adding user 'vagrant' to the group 'dialout' so it can access USB devices..."
 	usermod vagrant -a -G dialout
 	echo "DONE!"
 fi
-echo "MOMOROOT=$MOMOROOT" >> $HOME/.profile
+echo "export MOMOPATH=$MOMOPATH" >> $HOME/.profile
 mkdir -p $DOWNLOADCACHE
 
 download_file () {
@@ -46,42 +57,40 @@ download_file () {
 	fi
 }
 
-SCONSVERSION=2.3.3
-download_file "SCONS Installer" http://downloads.sourceforge.net/project/scons/scons/$SCONSVERSION/scons-$SCONSVERSION.tar.gz 63a74e1012ad9ad54950eafc69604fd9246e1a405c18acf936caa433c97cc0e2
-tar -xvzf scons-$SCONSVERSION.tar.gz
-cd scons-$SCONSVERSION
-python setup.py install
-echo "export PYTHONPATH=\".:/usr/local/lib/scons-$SCONSVERSION\"" >> $HOME/.profile
-cd ..
+if [ -n "$MOMO_DEV" ]; then
 
-XC8INSTALLER='xc8-v1.30-linux.run'
-download_file "XC8 Installer"	http://ww1.microchip.com/downloads/en/DeviceDoc/$XC8INSTALLER.tar 1ab6186eda45f774cfad1e2a0c7a4cbe449991b363634545c3eb005667925722
-tar -xvf $XC8INSTALLER.tar
-echo "Installing xc8 compiler..."
-./$XC8INSTALLER --mode unattended --netservername "" --prefix "/opt/microchip/xc8/v1.30"
-CODE=$?
-echo "export PATH=\"\$PATH:/opt/microchip/xc8/v1.30/bin\"" >> $HOME/.profile
-rm -f ./$XC8INSTALLER $XC8INSTALLER.tar
-if [ $CODE -ne 0 ]; then
-	die "Failed to install xc8, exiting!"
+	XC8VERSION=v1.34
+	XC8INSTALLER=xc8-$XC8VERSION-full-install-linux-installer.run
+	download_file "XC8 Installer"	http://ww1.microchip.com/downloads/en/DeviceDoc/$XC8INSTALLER ab0e36db27919fee2f41f95f345920e50fa3f32ba64f0accd2e041715ebdb231
+	echo "Installing xc8 compiler..."
+	chmod +x ./$XC8INSTALLER
+	./$XC8INSTALLER --mode unattended --netservername "" --prefix "/opt/microchip/xc8/$XC8VERSION"
+	CODE=$?
+	echo "export PATH=\"\$PATH:/opt/microchip/xc8/$XC8VERSION/bin\"" >> $HOME/.profile
+	rm -f ./$XC8INSTALLER $XC8INSTALLER.tar
+	if [ $CODE -ne 0 ]; then
+		die "Failed to install xc8, exiting!"
+	fi
+	echo "DONE!"
+
+	XC16VERSION=v1.24
+	XC16INSTALLER=xc16-$XC16VERSION-full-install-linux-installer.run
+	download_file "XC16 Installer" http://ww1.microchip.com/downloads/en/DeviceDoc/$XC16INSTALLER 22fd2d5cb7042e8d9aa7aef8c2e5a8c7a76bf3c55e0b5ff2dbd23450e227e023
+	echo "Installing xc16 compiler..."
+	chmod +x ./$XC16INSTALLER
+	./$XC16INSTALLER --mode unattended --netservername "" --prefix "/opt/microchip/xc16/$XC16VERSION"
+	CODE=$?
+	echo "export PATH=\"\$PATH:/opt/microchip/xc16/$XC16VERSION/bin\"" >> $HOME/.profile
+	rm -f ./$XC16INSTALLER ./$XC16INSTALLER.tar
+	if [ $CODE -ne 0 ]; then
+		die "Failed to install xc16, exiting!"
+	fi
+	echo "DONE!"
+
+	cd $HOME
+	$MOMOPATH/tools/automation/install_patched_gpsim.sh
 fi
-echo "DONE!"
 
-XC16INSTALLER=xc16-v1.21-linux-installer.run
-download_file "XC16 Installer" http://ww1.microchip.com/downloads/en/DeviceDoc/$XC16INSTALLER.tar 80a9fcc6e9e8b051266e06c1eff0ca078ebbc791a7d248dedd65f34a76d7735c
-tar -xvf $XC16INSTALLER.tar
-echo "Installing xc16 compiler..."
-./$XC16INSTALLER --mode unattended --netservername "" --prefix "/opt/microchip/xc16/v1.21"
-CODE=$?
-echo "export PATH=\"\$PATH:/opt/microchip/xc16/v1.21/bin\"" >> $HOME/.profile
-rm -f ./$XC16INSTALLER ./$XC16INSTALLER.tar
-if [ $CODE -ne 0 ]; then
-	die "Failed to install xc16, exiting!"
-fi
-echo "DONE!"
-
-echo "Adding MoMo tool bin ($MOMOROOT/tools/bin) to the path..."
-echo "export PATH=\"\$PATH:$MOMOROOT/tools/bin\"" >> $HOME/.profile
 echo "DONE!"
 
 exit 0
